+++ odoo/songs/install/data_all.py
@@ -3,6 +3,7 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html)
 
 import anthem
+from ..common import load_csv
 
 """ Data loaded in all modes
 The data loaded here will be loaded in the 'demo' and
@@ -11,5 +12,13 @@ The data loaded here will be loaded in the 'demo' and
 
 
 @anthem.log
+def import_analytic_account_project(ctx):
+    """ Importing analytic account (project) from csv """
+    load_csv(ctx, 'data/install/analytic_account_project.csv',
+             'account.analytic.account')
+
+
+@anthem.log
 def main(ctx):
     """ Loading data """
+    import_analytic_account_project(ctx)